+++ groupe87-demineur/nextjs-app/lib/constants.ts
@@ -1,0 +1,20 @@
+export const DIFFICULTY_PRESETS = {
+  beginner: { rows: 8, cols: 8, mines: 10 },
+  easy: { rows: 9, cols: 9, mines: 10 },
+  medium: { rows: 16, cols: 16, mines: 40 },
+  hard: { rows: 16, cols: 30, mines: 99 },
+  expert: { rows: 20, cols: 24, mines: 130 },
+  custom: { rows: 5, cols: 5, mines: 5 },
+}
+
+export const NUMBER_COLORS = [
+  "",
+  "text-blue-600",
+  "text-green-600",
+  "text-red-600",
+  "text-purple-800",
+  "text-yellow-800",
+  "text-teal-600",
+  "text-black",
+  "text-gray-600",
+]
