+++ groupe87-demineur/nextjs-app/lib/csp-solver.ts
@@ -1,14 +1,5 @@
 import type { CSPVariable, CSPConstraint } from "./types"
 
-/**
- * Solves a Minesweeper board using Constraint Satisfaction Problem (CSP) techniques
- *
- * @param board The current state of the board, with "?" for unknown cells and numbers for revealed cells
- * @param totalMines The total number of mines on the board
- * @param knownMines Optional array of known mine positions [row, col]
- * @param maxSolutions Maximum number of solutions to find
- * @returns Object containing solutions and recommendations
- */
 export function solveMinesweeperCSP(
   board: (number | string)[][],
   totalMines: number,
@@ -18,7 +9,6 @@ export function solveMinesweeperCSP(
   const rows = board.length
   const cols = board[0].length
 
-  // Create variables for each unknown cell
   const variables: CSPVariable[] = []
   const knownMinePositions = new Set(knownMines.map(([r, c]) => `${r},${c}`))
 
@@ -28,13 +18,12 @@ export function solveMinesweeperCSP(
         variables.push({
           row: r,
           col: c,
-          domain: [0, 1], // 0 = no mine, 1 = mine
+          domain: [0, 1],
         })
       }
     }
   }
 
-  // Create constraints based on revealed cells
   const constraints: CSPConstraint[] = []
 
   for (let r = 0; r < rows; r++) {
@@ -43,7 +32,6 @@ export function solveMinesweeperCSP(
         const adjacentVars: CSPVariable[] = []
         let knownAdjacentMines = 0
 
-        // Check all 8 adjacent cells
         for (let dr = -1; dr <= 1; dr++) {
           for (let dc = -1; dc <= 1; dc++) {
             if (dr === 0 && dc === 0) continue
@@ -74,20 +62,15 @@ export function solveMinesweeperCSP(
     }
   }
 
-  // Add constraint for total number of mines
   const totalMineConstraint: CSPConstraint = {
     variables: variables,
     sum: totalMines - knownMines.length,
   }
   constraints.push(totalMineConstraint)
 
-  // Find all solutions using backtracking
   const solutions = findAllSolutions(variables, constraints, maxSolutions)
-
-  // Calculate probabilities for each cell
   const probabilities = calculateProbabilities(rows, cols, solutions, knownMines)
 
-  // Generate recommendations
   const safeRecommendations: [number, number][] = []
   const mineRecommendations: [number, number][] = []
 
@@ -111,25 +94,17 @@ export function solveMinesweeperCSP(
   }
 }
 
-/**
- * Finds all valid solutions to the CSP using backtracking
- */
 function findAllSolutions(variables: CSPVariable[], constraints: CSPConstraint[], maxSolutions = 10): number[][][] {
   const solutions: number[][][] = []
   const assignment: Map<string, number> = new Map()
 
-  // Recursive backtracking function
   function backtrack(index: number) {
-    // If we've found enough solutions, stop searching
     if (solutions.length >= maxSolutions) {
       return
     }
 
-    // Base case: all variables assigned
     if (index === variables.length) {
-      // Check if all constraints are satisfied
       if (isAssignmentValid(assignment, constraints)) {
-        // Convert assignment to a solution board
         const solution = createSolutionBoard(variables, assignment)
         solutions.push(solution)
       }
@@ -139,11 +114,9 @@ function findAllSolutions(variables: CSPVariable[], constraints: CSPConstraint[]
     const variable = variables[index]
     const key = `${variable.row},${variable.col}`
 
-    // Try each value in the domain
     for (const value of variable.domain) {
       assignment.set(key, value)
 
-      // Check if partial assignment is valid
       if (isPartialAssignmentValid(assignment, constraints)) {
         backtrack(index + 1)
       }
@@ -152,15 +125,11 @@ function findAllSolutions(variables: CSPVariable[], constraints: CSPConstraint[]
     }
   }
 
-  // Start backtracking
   backtrack(0)
 
   return solutions
 }
 
-/**
- * Checks if a partial assignment is valid
- */
 function isPartialAssignmentValid(assignment: Map<string, number>, constraints: CSPConstraint[]): boolean {
   for (const constraint of constraints) {
     let sum = 0
@@ -176,17 +145,13 @@ function isPartialAssignmentValid(assignment: Map<string, number>, constraints: 
       }
     }
 
-    // If all variables are assigned, check if sum matches
     if (unassignedCount === 0) {
       if (sum !== constraint.sum) {
         return false
       }
     } else {
-      // Check if partial assignment is still feasible
       remainingSum -= sum
 
-      // If remaining sum is negative or greater than number of unassigned variables,
-      // this partial assignment cannot lead to a valid solution
       if (remainingSum < 0 || remainingSum > unassignedCount) {
         return false
       }
@@ -196,9 +161,6 @@ function isPartialAssignmentValid(assignment: Map<string, number>, constraints: 
   return true
 }
 
-/**
- * Checks if a complete assignment satisfies all constraints
- */
 function isAssignmentValid(assignment: Map<string, number>, constraints: CSPConstraint[]): boolean {
   for (const constraint of constraints) {
     let sum = 0
@@ -216,13 +178,9 @@ function isAssignmentValid(assignment: Map<string, number>, constraints: CSPCons
   return true
 }
 
-/**
- * Creates a solution board from an assignment
- */
 function createSolutionBoard(variables: CSPVariable[], assignment: Map<string, number>): number[][] {
   if (variables.length === 0) return []
 
-  // Find board dimensions
   let maxRow = 0
   let maxCol = 0
 
@@ -231,12 +189,10 @@ function createSolutionBoard(variables: CSPVariable[], assignment: Map<string, n
     maxCol = Math.max(maxCol, variable.col)
   }
 
-  // Create empty board
   const board = Array(maxRow + 1)
     .fill(null)
     .map(() => Array(maxCol + 1).fill(0))
 
-  // Fill in assigned values
   for (const variable of variables) {
     const key = `${variable.row},${variable.col}`
     board[variable.row][variable.col] = assignment.get(key) || 0
@@ -245,9 +201,6 @@ function createSolutionBoard(variables: CSPVariable[], assignment: Map<string, n
   return board
 }
 
-/**
- * Calculates the probability of a mine for each cell based on all solutions
- */
 function calculateProbabilities(
   rows: number,
   cols: number,
@@ -258,17 +211,15 @@ function calculateProbabilities(
     .fill(null)
     .map(() => Array(cols).fill(0))
 
-  // Mark known mines with 100% probability
   for (const [r, c] of knownMines) {
     probabilities[r][c] = 1
   }
 
   if (solutions.length === 0) return probabilities
 
-  // Calculate probabilities based on solutions
   for (let r = 0; r < rows; r++) {
     for (let c = 0; c < cols; c++) {
-      if (probabilities[r][c] === 1) continue // Skip known mines
+      if (probabilities[r][c] === 1) continue
 
       let mineCount = 0
